New package platform/kube/utils
