+++ pkg/platform/kube/apis/nuclio.io/v1beta1/types.go
@@ -17,9 +17,9 @@ limitations under the License.
 package v1beta1
 
 import (
-	"github.com/nuclio/nuclio/pkg/common"
 	"github.com/nuclio/nuclio/pkg/functionconfig"
 	"github.com/nuclio/nuclio/pkg/platform"
+	"github.com/nuclio/nuclio/pkg/platform/kube/utils"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
@@ -127,7 +127,7 @@ func (nf *NuclioFunction) GetComputedMaxReplicas() int32 {
 // EnrichNodeSelector enriches Spec.NodeSelector with platform and project NodeSelectors,
 // where function values take precedence over project values, and project values take precedence over platform values
 func (nf *NuclioFunction) EnrichNodeSelector(platformNodeSelector, projectNodeSelector map[string]string) {
-	nf.Status.EnrichedNodeSelector = common.MergeNodeSelector(nf.Spec.NodeSelector, projectNodeSelector, platformNodeSelector)
+	nf.Status.EnrichedNodeSelector = utils.MergeNodeSelector(nf.Spec.NodeSelector, projectNodeSelector, platformNodeSelector)
 }
 
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
